PC gefixt

+++ Simulator/Application/Services/Constants.cs
@@ -102,7 +102,7 @@ namespace Application.Services
         public static readonly short RBIF = 0x00;
         #endregion
         #region Quarzfrequenzen
-        public static readonly ObservableCollection<int> QuartzFrequencyList = new ObservableCollection<int> { 2000000, 4000000, 8000000, 16000000 };
+        public static readonly ObservableCollection<double> QuartzFrequencyList = new ObservableCollection<double> { 2000000, 4000000, 8000000, 16000000 };
         #endregion
     }
 }
